requirements.txt
pydantic>=2.13,<3
python-pptx>=1.0,<2
Pillow>=12,<13
PyMuPDF>=1.26,<2
graphifyy[mcp]>=0.4.23,<0.5
